renamed "core.vars_to_dim" into "core.vars2dim" renamed "core.levels" into "core.level" renamed "core.api" into "core.xarray"
