+++ xcube/core/timeseries.py
